fix: parsing value body post and add filter

+++ src/app/app/api/events/route.ts
@@ -1,24 +1,32 @@
 import { EventFormatting } from '@/database/event/createEvent';
 import { prisma } from '@/database/prisma/prisma';
 import { decrypt } from '@/lib/session';
+import { EventFormSchema, EventSearchFormSchema } from '@/schema/EventForm';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(req: NextRequest): Promise<NextResponse> {
 	try {
-		const from = req.nextUrl.searchParams.get('from');
-		const to = req.nextUrl.searchParams.get('to');
-		const limit = req.nextUrl.searchParams.get('limit');
-		let limitValue = 20;
-		if (limit != null) limitValue = Number(limit);
-		if (limitValue > 100) limitValue = 100;
-		if (from == null || to == null)
-			return new NextResponse('Error empty fields', {
+		const cookie = req.cookies.get('session');
+		const session = await decrypt(cookie?.value);
+
+		const fields = EventSearchFormSchema.safeParse({
+			from: req.nextUrl.searchParams.get('from'),
+			to: req.nextUrl.searchParams.get('to'),
+			limit: req.nextUrl.searchParams.get('limit'),
+			search: req.nextUrl.searchParams.get('search'),
+			club: req.nextUrl.searchParams.get('club'),
+			subscribe: req.nextUrl.searchParams.get('subscribe'),
+		});
+
+		if (!fields.success)
+			return new NextResponse(fields.error.message[0], {
 				status: 401,
 			});
-		const cookie = req.cookies.get('session');
-		await decrypt(cookie?.value);
+		if (fields.data.limit == null) fields.data.limit = 100;
+		else fields.data.limit > 0 && fields.data.limit <= 100 ? {} : (fields.data.limit = 100);
+
 		const value = await prisma.event.findMany({
-			take: limitValue,
+			take: fields.data.limit,
 			include: {
 				author: {
 					include: { memberships: true },
@@ -28,13 +36,27 @@ export async function GET(req: NextRequest): Promise<NextResponse> {
 			},
 			where: {
 				start_at: {
-					gte: new Date(from),
+					gte: fields.data.from,
 				},
 				end_at: {
-					lte: new Date(to),
+					lte: fields.data.to,
+				},
+				registered: {
+					...(fields.data.subscribe ? { user_id: session.user_id } : {}),
+				},
+				author: {
+					...(fields.data.club ? { full_name: fields.data.club } : {}),
+				},
+			},
+			orderBy: {
+				_relevance: {
+					fields: ['title'],
+					search: fields.data.search ? fields.data.search?.trim().split(/\s+/).join(' & ') : '',
+					sort: 'desc',
 				},
 			},
 		});
+
 		const events = await Promise.all(value.map(EventFormatting));
 		return NextResponse.json(events);
 	} catch (error: unknown) {
@@ -49,8 +71,15 @@ export async function DELETE(req: NextRequest): Promise<NextResponse> {
 	try {
 		const cookie = req.cookies.get('session');
 		await decrypt(cookie?.value);
+
 		const body = await req.json();
-		const row = await prisma.event.delete({
+
+		if (body.id == null)
+			return new NextResponse('Error, id not found.', {
+				status: 404,
+			});
+
+		await prisma.event.delete({
 			where: {
 				id: body.id,
 			},
@@ -59,7 +88,8 @@ export async function DELETE(req: NextRequest): Promise<NextResponse> {
 				image_album: true,
 			},
 		});
-		return NextResponse.json(row);
+
+		return NextResponse.json({success: true});
 	} catch (error: unknown) {
 		console.error(error);
 		return new NextResponse('Error, failed to delete event.', {
@@ -72,15 +102,31 @@ export async function POST(req: NextRequest): Promise<NextResponse> {
 	try {
 		const cookie = req.cookies.get('session');
 		const session = await decrypt(cookie?.value);
+
 		const body = await req.json();
-		const row = await prisma.event.create({
+
+		const fields = EventFormSchema.safeParse({
+			title: body.title,
+			description: body.description,
+			start_at: body.start_at,
+			end_at: body.end_at,
+			max_inscription: body.max_inscription,
+		});
+		if (!fields.success) {
+			console.error(fields.error.issues[0].message);
+			return new NextResponse(fields.error.issues[0].message, {
+				status: 401,
+			});
+		}
+
+		await prisma.event.create({
 			data: {
 				author_id: session.user_id,
-				title: body.title,
-				description: body.description,
-				max_inscription: body.max_inscription,
-				start_at: body.start_at,
-				end_at: body.end_at,
+				title: fields.data.title,
+				description: fields.data.description,
+				max_inscription: fields.data.max_inscription,
+				start_at: fields.data.start_at,
+				end_at: fields.data.end_at,
 			},
 			include: {
 				author: {
@@ -91,10 +137,11 @@ export async function POST(req: NextRequest): Promise<NextResponse> {
 				registered: true,
 			},
 		});
-		return NextResponse.json(EventFormatting(row));
+
+		return NextResponse.json({success: true});
 	} catch (error: unknown) {
 		console.error(error);
-		return new NextResponse('Error, failed to delete event.', {
+		return new NextResponse('Error, failed to create event.', {
 			status: 500,
 		});
 	}
